pip_requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
h11==0.12.0
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
python-decouple==3.4
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
